Renamed the file placetlattice.py -> lattice.py; import fixes
